requirements.txt
openai
python-dotenv
pathlib
numpy
sounddevice
pygame